packages/msa/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bio-analyze-msa"
version = "0.1.0a0"
description = "Multiple Sequence Alignment and Phylogenetic Tree module for bio-analyze."
readme = "README.md"
requires-python = ">=3.10, <3.15"
license = { text = "GPL-3.0" }
authors = [{ name = "qww" }]
dependencies = [
    "bio-analyze-core>=0.1.0a0",
    "biopython>=1.80",
    "agentyper>=0.1.12",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock"
]

[project.entry-points."bio_analyze.cli"]
msa = "bio_analyze_msa.cli:get_app"

[tool.hatch.build.targets.wheel]
packages = ["src/bio_analyze_msa"]
